Plane and movement
Here, I created an initial plane, which will be extended to be part of a tile function to create an infinite world. I included the Box struct from Lab1 to be a placeholder for the player. I implemented some basic movement with the WASD keys, and calculated boundaries so the box does not go past the borders of the plane.